Move graphql query to another file
