+++ packages/ssx-core/src/utils/utils.ts
@@ -14,9 +14,10 @@ import {
   SSXLensProfilesResponse,
   SSXPocketProviderNetworks,
   SSXRPCProvider,
-} from './types';
+} from '../types';
 import { ethers, getDefaultProvider } from 'ethers';
 import axios from 'axios';
+import { getProfilesQuery } from './queries';
 
 /**
  * @param rpc - SSXRPCProvider
@@ -159,97 +160,7 @@ export const ssxResolveLens = async (
       method: 'post',
       data: {
         operationName: 'Profiles',
-        query: `
-            query Profiles($addresses: [EthereumAddress!], $cursor: Cursor) {
-              profiles(request: { ownedBy: $addresses, limit: 10, cursor: $cursor }) {
-                items {
-                  id
-                  name
-                  bio
-                  attributes {
-                    displayType
-                    traitType
-                    key
-                    value
-                  }
-                  followNftAddress
-                  metadata
-                  isDefault
-                  picture {
-                    ... on NftImage {
-                      contractAddress
-                      tokenId
-                      uri
-                      verified
-                    }
-                    ... on MediaSet {
-                      original {
-                        url
-                        mimeType
-                      }
-                    }
-                    __typename
-                  }
-                  handle
-                  coverPicture {
-                    ... on NftImage {
-                      contractAddress
-                      tokenId
-                      uri
-                      verified
-                    }
-                    ... on MediaSet {
-                      original {
-                        url
-                        mimeType
-                      }
-                    }
-                    __typename
-                  }
-                  ownedBy
-                  dispatcher {
-                    address
-                    canUseRelay
-                  }
-                  stats {
-                    totalFollowers
-                    totalFollowing
-                    totalPosts
-                    totalComments
-                    totalMirrors
-                    totalPublications
-                    totalCollects
-                  }
-                  followModule {
-                    ... on FeeFollowModuleSettings {
-                      type
-                      amount {
-                        asset {
-                          symbol
-                          name
-                          decimals
-                          address
-                        }
-                        value
-                      }
-                      recipient
-                    }
-                    ... on ProfileFollowModuleSettings {
-                    type
-                    }
-                    ... on RevertFollowModuleSettings {
-                    type
-                    }
-                  }
-                }
-                pageInfo {
-                  prev
-                  next
-                  totalCount
-                }
-              }
-            }
-        `,
+        query: getProfilesQuery,
         variables: {
           addresses: [address],
           cursor: pageCursor
